Complete rewrite
The previous commits were a complete mess the algorithm seemed to have
turned out to be impossible to implement. I also made a huge mistake of
not writing this in proper OOP practices and instead just used a bunch
of static functions. The entire project is now beyond repair and the
only way to fix it is to start over. Therefore, all files except
Utilities.cs and its tests are deleted.